refactor: rename directory removing Block suffix
