+++ src/Campaigns/Blocks/CampaignDonors/edit.tsx
@@ -43,7 +43,7 @@ export default function Edit({
     return (
         <div {...blockProps}>
             <CampaignSelector attributes={attributes} setAttributes={setAttributes}>
-                <ServerSideRender block="givewp/campaign-donors-block" attributes={attributes} />
+                <ServerSideRender block="givewp/campaign-donors" attributes={attributes} />
             </CampaignSelector>
 
             {hasResolved && campaign?.id && (
