+++ src/Campaigns/Blocks/CampaignTitle/block.json
@@ -1,7 +1,7 @@
 {
     "$schema": "https://schemas.wp.org/trunk/block.json",
     "apiVersion": 3,
-    "name": "givewp/campaign-title-block",
+    "name": "givewp/campaign-title",
     "version": "1.0.0",
     "title": "Campaign Title",
     "category": "give",
