+++ src/Campaigns/Blocks/blocks.ts
@@ -1,10 +1,10 @@
-import * as campaignTitleBlock from './CampaignTitleBlock';
 import * as campaignCover from './CampaignCover';
 import * as campaignDonateButton from './DonateButton';
 import * as campaignDonors from './CampaignDonors';
+import * as campaignTitle from './CampaignTitle';
 
 const getAllBlocks = () => {
-    return [campaignCover, campaignDonateButton, campaignDonors, campaignTitleBlock];
+    return [campaignCover, campaignDonateButton, campaignDonors, campaignTitle];
 };
 
 getAllBlocks().forEach((block) => {
